% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulationResultClass.R
\name{new_SimulationResult}
\alias{new_SimulationResult}
\title{Constructor for the SimulationResult class.}
\usage{
new_SimulationResult(resultJSON)
}
\arguments{
\item{resultJSON}{the output of a simulation call most likely
deserialized from a JSON string}
}
\value{
an S3 SimulationResult instance
}
\description{
This class converts the  result of a simulation call
into a SimulationResult instance
}
\details{
The constructor automatically detects whether the
simulation results need to be aggregated.
}
